edit 5215

+++ CMSGuias/apps/ventas/views.py
@@ -882,47 +882,86 @@ class SectorManage(JSONResponseMixin, View):
                             model_id=request.POST.get('models'),
                             flag=True
                         )
-                    meter = MetProject.objects.get(
+
+                    obj.brand_id = request.POST.get('brand')
+                    obj.model_id = request.POST.get('model')
+                    pending = 0
+                    try:
+                        meter = MetProject.objects.get(
                             proyecto_id=kwargs['pro'],
                             subproyecto_id=kwargs['sub'] if kwargs['sub'] != unicode(None) else None,
                             sector_id=kwargs['sec'],
-                            materials_id=request.POST.get('materials'),
+                            materiales_id=request.POST.get('materials'),
                             brand_id=request.POST.get('brands'),
-                            model_id=request.POST.get('models'),
-                            flag=True
+                            model_id=request.POST.get('models')
                         )
-                    obj.brand_id = request.POST.get('brand')
-                    obj.model_id = request.POST.get('model')
+
+                        pending = float(meter.quantityorder)
+
+                    except ObjectDoesNotExist:
+                        pending = 0
 
                     quantity = float(request.POST.get('quantity'))
                     orders = obj.quantityorders
-                    if quantity > 0 and quantity > obj.quantity: # tag 0 or 1
-                        if orders == 0:
-                            obj.quantityorders = (quantity - obj.quantity)
-                            obj.tag = '1'
-                            print 'nc mayor tag 1'
-                        elif orders == obj.quantity:
+                    print 'pending proccess', pending
+                    if quantity > obj.quantity:
+                        if pending > 0 or pending == 0:
+                            if orders == obj.quantity:
+                                obj.tag = '0'
+                            else:
+                                obj.tag = '1'
+                            obj.quantityorders = (pending + (quantity - meter.cantidad))
+                        if pending == quantity:
                             obj.quantityorders = quantity
                             obj.tag = '0'
-                            print 'orders = ca tag 0'
-                    elif quantity > 0 and quantity < obj.quantity: # 1 or 2
-                        if orders == 0:
-                            obj.tag = '2'
-                            print 'tag 2'
-                        if orders > 0 and orders < quantity:
-                            obj.quantityorders = ((obj.quantity - quantity) - obj.quantityorders)
-                            print 'nc < orders tag 1'
-                            if obj.quantityorders > 0:
-                                obj.tag = '1'
-                            else:
+                    elif quantity < obj.quantity:
+                        if pending == 0:
+                            tag = '2'
+                            obj.quantityorders = pending
+                        if pending >= quantity:
+                            obj.quantityorders = quantity
+                            obj.tag = '0'
+                        if pending < quantity:
+                            obj.quantityorders = (pending - (obj.quantity - quantity))
+                            if (pending - (obj.quantity - quantity)) == 0:
                                 obj.tag = '2'
-                        if orders > 0 and orders > quantity:
-                            orde = obj.quantity - orders
-                            obj.quantityorders = ((orders - quantity) + orde)
-                            obj.tag = '1'
-                            print 'orders > nc tag = 1'
+                            else:
+                                obj.tag = '1'
                     else:
-                        obj.tag = '0'
+                        if pending == 0 and quantity == meter.cantidad:
+                            obj.tag = '2'
+                        else:
+                            obj.tag = '0'
+                    # orders = obj.quantityorders
+                    # if quantity > 0 and quantity > obj.quantity: # tag 0 or 1
+                    #     if orders == 0:
+                    #         obj.quantityorders = (quantity - obj.quantity)
+                    #         obj.tag = '1'
+                    #         print 'nc mayor tag 1'
+                    #     elif orders == obj.quantity:
+                    #         obj.quantityorders = quantity
+                    #         obj.tag = '0'
+                    #         print 'orders = ca tag 0'
+                    # elif quantity > 0 and quantity < obj.quantity: # 1 or 2
+                    #     if orders == 0:
+                    #         obj.tag = '2'
+                    #         print 'tag 2'
+                    #     if orders > 0 and orders < quantity:
+                    #         obj.quantityorders = ((obj.quantity - quantity) - obj.quantityorders)
+                    #         print 'nc < orders tag 1'
+                    #         if obj.quantityorders > 0:
+                    #             obj.tag = '1'
+                    #         else:
+                    #             obj.tag = '2'
+                    #     if orders > 0 and orders > quantity:
+                    #         orde = obj.quantity - orders
+                    #         obj.quantityorders = ((orders - quantity) + orde)
+                    #         obj.tag = '1'
+                    #         print 'orders > nc tag = 1'
+                    # else:
+                    #     obj.tag = '0'
+                    #
+                    #
                     # if obj.tag != '0':
                     #     if obj.quantity < float(request.POST.get('quantity')):
                     #         if obj.tag == '1':
@@ -960,6 +999,15 @@ class SectorManage(JSONResponseMixin, View):
                         materials_id=request.POST.get('materials'),
                         brand_id=request.POST.get('brand'),
                         model_id=request.POST.get('model'),
+                        flag=True
+                    )
+                    obj.delete()
+                    context['status'] = True
+                if 'deleteallupdatemeter' in request.POST:
+                    obj = UpdateMetProject.objects.filter(
+                        proyecto_id=kwargs['pro'],
+                        subproyecto_id=kwargs['sub'] if kwargs['sub'] != unicode(None) else None,
+                        sector_id=kwargs['sec'],
                         flag=True
                     )
                     obj.delete()
